refactor: Move TransferManagerConfig to own file and Add Parallel Upload/Download Configs to jobs
